test: consolidate MC search helpers into one file per PR #1048 feedback
Per PR #1048 review comments ("Why is this a separate file? This
separation of concern, concerns me." / "don't generate big test files
for functions like this" / "Does this stuff need to be in it's own
file?"), fold the 4-file tests/common/multicluster_search/ package
plus its 3 helper-test files into one module at
tests/common/search/multicluster_search_helper.py.

Matches the existing per-topology pattern in tests/common/search/
(replicaset_search_helper.py, sharded_search_helper.py). Drops the
unit tests for test helpers — those add maintenance surface for code
the e2e itself exercises.
